made code separation
