+++ app/src/main/java/com/nameisjayant/noteappcompose/features/ui/screens/LoadingBar.kt
@@ -1,0 +1,11 @@
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.window.Dialog
+import com.nameisjayant.noteappcompose.ui.theme.Red
+
+@Composable
+fun LoadingBar() {
+    Dialog(onDismissRequest = {}) {
+        CircularProgressIndicator(color = Red)
+    }
+}
